+++ alkymi/__init__.py
@@ -2,9 +2,10 @@ from .config import AlkymiConfig  # NOQA
 from .lab import Lab  # NOQA
 from .logging import log  # NOQA
 from .decorators import recipe, foreach  # NOQA
-from .recipes import glob_files, file, args, kwargs  # NOQA
+from .recipes import glob_files, file, arg  # NOQA
 from .utils import call  # NOQA
 from .types import Status  # NOQA
+from . import checksums  # NOQA
 
 # Define version
 from .version import VERSION, __version__  # NOQA