+++ alkymi/lab.py
@@ -1,10 +1,12 @@
 import argparse
 import logging
-from typing import Dict, Union, Any, List
+import sys
+from typing import Dict, Union, Any, List, Iterable, Optional, TextIO
 
 from .alkymi import compute_status_with_cache, Status
 from .logging import log
 from .recipe import Recipe
+from .recipes import Arg
 
 
 class Lab:
@@ -21,7 +23,8 @@ class Lab:
         :param name: The name of the Lab
         """
         self._name = name
-        self._recipes = []  # type: List[Recipe]
+        self._recipes: List[Recipe] = []
+        self._args: Dict[str, Arg] = {}
 
     def add_recipe(self, recipe: Recipe) -> Recipe:
         """
@@ -42,6 +45,14 @@ class Lab:
         """
         for recipe in recipes:
             self.add_recipe(recipe)
+
+    def register_arg(self, arg: Arg) -> None:
+        """
+        Register an argument with the Lab (this will make the argument settable through the CLI)
+
+        :param arg: The argument to register
+        """
+        self._args[arg.name] = arg
 
     def brew(self, target_recipe: Union[Recipe, str]) -> Any:
         """
@@ -76,16 +87,37 @@ class Lab:
         """
         return self._recipes
 
+    @property
+    def args(self) -> Dict[str, Arg]:
+        """
+        :return: The list of args registered with this Lab
+        """
+        return self._args
+
     def _build_full_status(self) -> Dict[Recipe, Status]:
         """
         Compute statuses for all recipes (and dependent recipes) in this Lab
 
         :return: The statuses as a dictionary
         """
-        status = {}  # type: Dict[Recipe, Status]
+        status: Dict[Recipe, Status] = {}
         for recipe in self._recipes:
             compute_status_with_cache(recipe, status)
         return status
+
+    def _add_user_args_(self, parser: argparse.ArgumentParser) -> None:
+        """
+        Adds user provided arguments to an ArgumentParser instance
+
+        :param parser: The parser to add the user-provided arguments to
+        """
+        for arg_name, arg in self._args.items():
+            # For iterables (e.g. lists), the "type" keyword is actually the type of elements in the iterable
+            if issubclass(arg.type, Iterable) and not arg.type == str:
+                subtype = arg.subtype if arg.subtype is not None else str
+                parser.add_argument("--{}".format(arg_name), type=subtype, nargs="*")
+            else:
+                parser.add_argument("--{}".format(arg_name), type=arg.type)
 
     def __repr__(self) -> str:
         """
@@ -97,11 +129,21 @@ class Lab:
             state += '\n\t{} - {}'.format(recipe.name, status[recipe])
         return '{} lab with recipes:{}'.format(self.name, state)
 
-    def open(self) -> None:
+    def open(self, args: Optional[List[str]] = None, stream: TextIO = sys.stderr) -> None:
         """
         Runs the command line interface for this Lab by parsing command line arguments and carrying out the designated
         command
+
+        :param args: The input arguments to use - will default to system args
+        :param stream: The stream to print output to
         """
+        if len(self.recipes) == 0:
+            raise RuntimeError("No recipes added to lab - CLI is useless")
+
+        # Use system args if nothing has been provided
+        if args is None:
+            args = sys.argv[1:]
+
         # Create the top-level parser
         parser = argparse.ArgumentParser('CLI for {}'.format(self._name))
         parser.add_argument("-v", "--verbose", action="store_true", help="Turn on verbose logging")
@@ -109,25 +151,33 @@ class Lab:
         subparsers = parser.add_subparsers(help='sub-command help', dest='subparser_name')
 
         # Create the parser for the "status" command
-        subparsers.add_parser('status', help='Prints the detailed status of the lab')
+        status_parser = subparsers.add_parser('status', help='Prints the detailed status of the lab')
+        self._add_user_args_(status_parser)
 
         # Create the parser for the "brew" command
         brew_parser = subparsers.add_parser('brew', help='Brew the selected recipe')
         brew_parser.add_argument('recipe', choices=[recipe.name for recipe in self._recipes], nargs="+",
                                  help='Recipe(s) to brew')
+        self._add_user_args_(brew_parser)
 
-        args = parser.parse_args()
-        log.addHandler(logging.StreamHandler())
-        if args.verbose:
+        parsed_args = parser.parse_args(args)
+        log.addHandler(logging.StreamHandler(stream))
+        if parsed_args.verbose:
             log.setLevel(logging.DEBUG)
         else:
             log.setLevel(logging.INFO)
 
-        if args.subparser_name == 'status':
-            print(self)
-        elif args.subparser_name == 'brew':
-            for recipe in args.recipe:
+        # Set arguments if supplied
+        for arg_name, arg in self._args.items():
+            provided_val = getattr(parsed_args, arg_name, None)
+            if provided_val is not None:
+                arg.set(provided_val)
+
+        if parsed_args.subparser_name == 'status':
+            print(self, file=stream)
+        elif parsed_args.subparser_name == 'brew':
+            for recipe in parsed_args.recipe:
                 self.brew(recipe)
         else:
             # No recognized command provided - print help
-            parser.print_help()
+            parser.print_help(file=stream)