+++ tests/test_foreach.py
@@ -9,7 +9,7 @@ from alkymi.alkymi import Status
 import alkymi as alk
 
 # We use these globals to avoid altering the hashes of bound functions when any of these change
-execution_counts = {}  # type: Dict[Union[Path, str], int]
+execution_counts: Dict[Union[Path, str], int] = {}
 f1 = Path()
 f2 = Path()
 f3 = Path()
@@ -39,67 +39,67 @@ def test_execution(caplog, tmpdir):
         assert execution_counts[f2.stem] == counts[4]
         assert execution_counts[f3.stem] == counts[5]
 
-    args = alkymi.recipes.args([f1], name="args")
+    arg = alkymi.recipes.arg([f1], name="arg")
 
-    @alk.foreach(args)
-    def read_file(path: Path) -> str:
+    @alk.foreach(arg)
+    def file_contents(path: Path) -> str:
         execution_counts[path] += 1
         with path.open('r') as f:
             return f.read()
 
-    @alk.recipe(ingredients=[read_file])
+    @alk.recipe()
     def to_dict(file_contents: List[str]) -> Dict[str, str]:
         return {f: f for f in file_contents}
 
     # This is used to check later whether changing an ingredient will correctly change everything
-    extra_count_arg = alkymi.recipes.args(0, name="extra_count_arg")
+    extra_count = alkymi.recipes.arg(0, name="extra_count_arg")
 
-    @alk.foreach(to_dict, ingredients=[extra_count_arg])
+    @alk.foreach(to_dict)
     def change_count(file_content: str, extra_count: int) -> str:
         execution_counts[file_content] += 1 + extra_count
         return file_content
 
-    assert read_file.status() == Status.NotEvaluatedYet
+    assert file_contents.status() == Status.NotEvaluatedYet
     assert change_count.status() == Status.NotEvaluatedYet
     change_count.brew()
-    assert read_file.status() == Status.Ok
+    assert file_contents.status() == Status.Ok
     assert change_count.status() == Status.Ok
     _check_counts((1, 0, 0, 1, 0, 0))
 
-    args.set_args([f1, f2, f3])
-    assert read_file.status() == Status.MappedInputsDirty
+    arg.set([f1, f2, f3])
+    assert file_contents.status() == Status.MappedInputsDirty
     change_count.brew()
-    assert read_file.status() == Status.Ok
+    assert file_contents.status() == Status.Ok
     _check_counts((1, 1, 1, 1, 1, 1))
 
-    args.set_args([f3, f2])
+    arg.set([f3, f2])
     change_count.brew()
     _check_counts((1, 1, 1, 1, 1, 1))
 
-    args.set_args([f1])
+    arg.set([f1])
     change_count.brew()
     _check_counts((2, 1, 1, 2, 1, 1))
 
-    args.set_args([])
+    arg.set([])
     change_count.brew()
     _check_counts((2, 1, 1, 2, 1, 1))
 
     # Test that changing an ingredient forces reevaluation of all foreach inputs
-    args.set_args([f1, f2, f3])
+    arg.set([f1, f2, f3])
     assert change_count.status() == Status.MappedInputsDirty
     change_count.brew()
     assert change_count.status() == Status.Ok
     _check_counts((3, 2, 2, 3, 2, 2))
 
     # This should cause a reevaluation of everything (+1 to all counts) and then add 10 from this arg
-    extra_count_arg.set_args(10)
+    extra_count.set(10)
     assert change_count.status() == Status.IngredientDirty
     change_count.brew()
     assert change_count.status() == Status.Ok
     _check_counts((3, 2, 2, 14, 13, 13))
 
 
-execution_counts_list = []  # type: List[int]
+execution_counts_list: List[int] = []
 
 
 def test_lists(caplog):
@@ -111,7 +111,7 @@ def test_lists(caplog):
 
     global execution_counts_list
     execution_counts_list = [0] * 5
-    args = alkymi.recipes.args([0], name="args")
+    arg = alkymi.recipes.arg([0], name="args")
 
     def _check_counts(counts: Tuple[int, int, int, int, int]):
         assert execution_counts_list[0] == counts[0]
@@ -120,7 +120,7 @@ def test_lists(caplog):
         assert execution_counts_list[3] == counts[3]
         assert execution_counts_list[4] == counts[4]
 
-    @alk.foreach(args)
+    @alk.foreach(arg)
     def record_execution(idx: int) -> int:
         execution_counts_list[idx] += 1
         return execution_counts_list[idx]
@@ -134,21 +134,72 @@ def test_lists(caplog):
     _check_counts((1, 0, 0, 0, 0))
 
     # Also brew ids 1 and 2
-    args.set_args([0, 1, 2])
+    arg.set([0, 1, 2])
     record_execution.brew()
     _check_counts((1, 1, 1, 0, 0))
 
     # Now only 3 and 4 - all items should now have been run once
-    args.set_args([3, 4])
+    arg.set([3, 4])
     record_execution.brew()
     _check_counts((1, 1, 1, 1, 1))
 
     # Because last execution was 3 and 4, asking for 0 and 1 should cause reevaluation
-    args.set_args([0, 1])
+    arg.set([0, 1])
     record_execution.brew()
     _check_counts((2, 2, 1, 1, 1))
 
     # Switching the order should not change anything
-    args.set_args([1, 0])
+    arg.set([1, 0])
     record_execution.brew()
     _check_counts((2, 2, 1, 1, 1))
+
+
+def test_bound_function_changed(caplog):
+    """
+    Test that changing the bound function of a ForeachRecipe causes a full re-evaluation
+    """
+    caplog.set_level(logging.DEBUG)
+    AlkymiConfig.get().cache = False
+
+    inputs = alk.arg([1, 2, 3, 4, 5], "inputs")
+    power = 2  # Note that power is captured in 'raised_inputs' - changing it changes the function hash
+
+    @alk.foreach(inputs)
+    def raised_inputs(val: int) -> int:
+        return val ** power
+
+    squared_inputs = raised_inputs.brew()
+    assert squared_inputs == [1, 4, 9, 16, 25]
+
+    # Changing power to something else should cause full re-evaluation due to the bound function changing
+    power = 3
+    cubed_inputs = raised_inputs.brew()
+    assert cubed_inputs == [1, 8, 27, 64, 125]
+
+
+def test_invalid_foreach_outputs(caplog, tmpdir):
+    """
+    Test that outputs of a ForeachRecipe will be re-evaluated on request if invalid (i.e. an external file pointed to by
+    a Path has been deleted)
+    """
+    tmpdir = Path(str(tmpdir))
+    caplog.set_level(logging.DEBUG)
+    AlkymiConfig.get().cache = False
+
+    inputs = alk.arg([1, 2, 3], "inputs")
+
+    @alk.foreach(inputs)
+    def files_with_values(val: int) -> Path:
+        f = Path(tmpdir) / (str(val) + ".txt")
+        f.write_text(str(val))
+        return f
+
+    # Initial evaluation should create all files
+    files: List[Path] = files_with_values.brew()
+    assert all([file.is_file() for file in files])
+
+    # If one of output files is deleted, a subsequent brew() should recreate it
+    files[0].unlink()
+    assert files_with_values.status() == Status.OutputsInvalid
+    files: List[Path] = files_with_values.brew()
+    assert all([file.is_file() for file in files])